backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile /Users/danielwaltrip/all-files/projects/code/frac-vizzy/backend/requirements.in
#
asgiref==3.5.2
    # via django
django==4.1.1
    # via
    #   -r /Users/danielwaltrip/all-files/projects/code/frac-vizzy/backend/requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.13.0
    # via -r /Users/danielwaltrip/all-files/projects/code/frac-vizzy/backend/requirements.in
djangorestframework==3.13.1
    # via -r /Users/danielwaltrip/all-files/projects/code/frac-vizzy/backend/requirements.in
pillow==9.2.0
    # via -r /Users/danielwaltrip/all-files/projects/code/frac-vizzy/backend/requirements.in
psycopg2==2.9.3
    # via -r /Users/danielwaltrip/all-files/projects/code/frac-vizzy/backend/requirements.in
pytz==2022.2.1
    # via djangorestframework
sqlparse==0.4.2
    # via django
